# internal/ui/web/openapi.yaml
openapi: 3.0.3
info:
  title: Velix API
  description: |
    API WhatsApp multi-instância. Gerencie instâncias, envie mensagens (texto, mídia, reações),
    receba eventos em tempo real via webhooks ou WebSocket e agende mensagens.
  version: "1.4.0"
  contact:
    name: Velix Support
  x-logo:
    url: /admin/dist/velix-logo.svg
    altText: Velix
servers:
  - url: /v1
    description: Current server
security:
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /instances/{instanceID}/messages/text:
    post:
      tags: [Messages]
      summary: Send or schedule a text message
      description: |
        If `send_at` is provided and in the future, the message is scheduled and
        returns `202 Accepted`. Otherwise it is sent immediately and returns `201 Created`.
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTextRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '202':
          description: Message scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /instances/{instanceID}/messages/media:
    post:
      tags: [Messages]
      summary: Send or schedule a media message
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMediaRequest'
      responses:
        '201':
          description: Sent
        '202':
          description: Scheduled
  /instances/{instanceID}/messages/reaction:
    post:
      tags: [Messages]
      summary: Send a reaction to a message
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendReactionRequest'
      responses:
        '200':
          description: OK
  /instances/{instanceID}/messages/read:
    post:
      tags: [Messages]
      summary: Mark messages as read
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkReadRequest'
      responses:
        '200':
          description: OK
  /instances/{instanceID}/messages/batch:
    post:
      tags: [Messages]
      summary: Send up to 100 text messages in a single call
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTextRequest'
      responses:
        '207':
          description: Multi-status — one result per message in request order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchResultItem'
  /instances/{instanceID}/messages/location:
    post:
      tags: [Messages]
      summary: Send a location pin
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendLocationRequest'
      responses:
        '201':
          description: Sent
  /instances/{instanceID}/messages/poll:
    post:
      tags: [Messages]
      summary: Send an interactive poll
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPollRequest'
      responses:
        '201':
          description: Sent
  /instances/{instanceID}/messages/contact:
    post:
      tags: [Messages]
      summary: Send one or more contact vCards
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendContactRequest'
      responses:
        '201':
          description: Sent
  /instances/{instanceID}/messages/edit:
    post:
      tags: [Messages]
      summary: Edit a previously sent message
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /instances/{instanceID}/messages/disappearing:
    post:
      tags: [Messages]
      summary: Set the disappearing message timer for a chat
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDisappearingTimerRequest'
      responses:
        '200':
          description: OK
  /instances/{instanceID}/messages/status:
    post:
      tags: [Messages]
      summary: Post a WhatsApp Status (Story)
      description: |
        Posts a status update visible to all contacts. Supports text (with optional
        background color and font style), image, and video.
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendStatusRequest'
      responses:
        '201':
          description: Status posted
  /instances/{instanceID}/messages:
    get:
      tags: [Messages]
      summary: List messages for a chat
      description: |
        Returns stored messages for a contact or group chat, newest first by send time.
        Includes messages received live and messages imported by history sync
        (`content.history = true`). To import older messages, see `POST /messages/history`.
      parameters:
        - $ref: '#/components/parameters/instanceID'
        - name: chat
          in: query
          required: true
          schema:
            type: string
          description: Chat JID (e.g. 5511999999999@s.whatsapp.net or 120363012345678901@g.us)
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /instances/{instanceID}/messages/history:
    post:
      tags: [Messages]
      summary: Request older messages for a chat from WhatsApp
      description: |
        Asks the primary device for up to `count` messages older than the oldest
        message already stored for the chat. The request is asynchronous: the phone
        replies with a history sync a few seconds later, the messages are stored, and
        they then appear in `GET /messages?chat=...`. Call repeatedly to page further back.

        Requires at least one stored message in the chat (the anchor) and the phone
        to be online. WhatsApp may return fewer messages than requested.
      parameters:
        - $ref: '#/components/parameters/instanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestHistoryRequest'
      responses:
        '202':
          description: Request sent to the phone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestHistoryResponse'
        '422':
          description: Missing/invalid chat or count, or no stored message to anchor the request
  /instances/{instanceID}/messages/scheduled:
    get:
      tags: [Messages]
      summary: List scheduled and cancelled messages for an instance
      parameters:
        - $ref: '#/components/parameters/instanceID'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: OK
  /instances/{instanceID}/messages/{msgID}:
    delete:
      tags: [Messages]
      summary: Revoke (delete for everyone) a sent message
      parameters:
        - $ref: '#/components/parameters/instanceID'
        - $ref: '#/components/parameters/msgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to:
                  type: string
                  description: Chat JID where the message lives
      responses:
        '204':
          description: Revoked
  /instances/{instanceID}/messages/{msgID}/schedule:
    delete:
      tags: [Messages]
      summary: Cancel a scheduled message
      parameters:
        - $ref: '#/components/parameters/instanceID'
        - $ref: '#/components/parameters/msgID'
      responses:
        '204':
          description: Cancelled

  # ─── Contacts ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    instanceID:
      name: instanceID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    msgID:
      name: msgID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
...
    SendTextRequest:
      type: object
      required: [to, text]
      properties:
        to:
          type: string
          description: Phone number (e.g. 5511999999999) or full JID
        text:
          type: string
        quoted_id:
          type: string
          description: WhatsApp message ID to quote/reply to
        send_at:
          type: string
          format: date-time
          description: Future timestamp to schedule delivery (RFC3339)
    SendMediaRequest:
      type: object
      required: [to, type, data, mime_type]
      properties:
        to:
          type: string
        type:
          type: string
          enum: [image, video, audio, document, sticker]
        data:
          type: string
          format: byte
          description: Base64-encoded file bytes
        mime_type:
          type: string
        file_name:
          type: string
        caption:
          type: string
        send_at:
          type: string
          format: date-time
    SendReactionRequest:
      type: object
      required: [to, message_id]
      properties:
        to:
          type: string
        message_id:
          type: string
        reaction:
          type: string
          description: Emoji character, or empty string to remove the reaction
    MarkReadRequest:
      type: object
      required: [chat, message_ids]
      properties:
        chat:
          type: string
        message_ids:
          type: array
          items:
            type: string
    BatchTextRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          maxItems: 100
          items:
            type: object
            required: [to, text]
            properties:
              to:
                type: string
              text:
                type: string
              send_at:
                type: string
                format: date-time
    BatchResultItem:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/Message'
        error:
          type: string
    SendLocationRequest:
      type: object
      required: [to, lat, lng]
      properties:
        to:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        name:
          type: string
        address:
          type: string
    SendPollRequest:
      type: object
      required: [to, question, options]
      properties:
        to:
          type: string
        question:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 12
          items:
            type: string
        multi_select:
          type: boolean
          default: false
    SendContactRequest:
      type: object
      required: [to, contacts]
      properties:
        to:
          type: string
        contacts:
          type: array
          items:
            type: object
            required: [name, phone]
            properties:
              name:
                type: string
              phone:
                type: string
    EditMessageRequest:
      type: object
      required: [to, message_id, text]
      properties:
        to:
          type: string
          description: Chat JID where the original message was sent
        message_id:
          type: string
          description: WhatsApp message ID of the message to edit
        text:
          type: string
          description: New text content
    SetDisappearingTimerRequest:
      type: object
      required: [chat, seconds]
      properties:
        chat:
          type: string
          description: JID of the chat (user or group)
        seconds:
          type: integer
          description: |
            Timer duration in seconds.
            - `0` — off
            - `86400` — 24 hours
            - `604800` — 7 days
            - `7776000` — 90 days
    RequestHistoryRequest:
      type: object
      required: [chat]
      properties:
        chat:
          type: string
          description: JID of the chat (user or group)
          example: 120363012345678901@g.us
        count:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
          description: Number of older messages to request
    RequestHistoryResponse:
      type: object
      properties:
        status:
          type: string
          example: requested
        chat:
          type: string
        count:
          type: integer
        anchor_message_id:
          type: string
          description: WhatsApp ID of the oldest stored message the request pages back from
        anchor_timestamp:
          type: string
          format: date-time
    SendStatusRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [text, image, video]
        caption:
          type: string
          description: Text content (for type=text) or caption (for image/video)
        background_color:
          type: string
          description: Hex color for text status background (e.g. "#000000")
        font_type:
          type: integer
          description: |
            Font style for text status:
            - `0` — SansSerif (default)
            - `1` — Serif
            - `2` — Norican
            - `3` — Bryndan Write
            - `4` — Bebas Neue
            - `5` — Oswald
        data:
          type: string
          format: byte
          description: Base64-encoded image or video bytes (required for image/video)
        mime_type:
          type: string
          description: MIME type (required for image/video, e.g. image/jpeg, video/mp4)
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        instance_id:
          type: string
        whatsapp_message_id:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [pending, scheduled, queued, sent, delivered, read, failed, revoked, cancelled]
        from_jid:
          type: string
        to_jid:
          type: string
        chat_jid:
          type: string
        is_group:
          type: boolean
        type:
          type: string
          enum: [text, image, video, audio, document, sticker, reaction, location, poll, contact]
        content:
          type: object
        scheduled_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...